# Improve aperture performance