… tests/baseline_for_apertures/circular/pupil_large.fits.gz
… tests/baseline_for_apertures/circular/pupil_small.fits.gz
… tests/baseline_for_apertures/rectangular/pupil_elongated.fits.gz
… tests/baseline_for_apertures/rectangular/pupil_square.fits.gz