This PR improves the speed of spiders and hexagonal segmented apertures by ~4x. This PR also adds tests for the circular and rectangular apertures.

Method
---
For spiders, I wrote specialized code for separated Cartesian grids, the most used grid in HCIPy. Most calculations can be performed much quicker because of the reduced array sizes,

For segmented apertures, I introduced a new keyword for Field generators; `return_with_mask`. This keyword, if it is `True` (default `False`), the function does not return a Field, but rather a `ndarray, mask` tuple. The mask can either be a tuple or a ndarray, to be used to mask the shaped or non-shaped version of a zero field with the original grid. This keyword allows for much more simplified additions to existing fields, which the `make_segmented_aperture()` function exploits to get higher performance.

This new keyword is not required for any Field generators, and is only used if it's available (as probed by the `inspect.signature()` function.

Benchmarks
---
Setup:
```python
pupil_grid = make_pupil_grid(512)

circle = circular_aperture(1)
rectangle = rectangular_aperture([1, 1])
hexagon = hexagonal_aperture(1)
spider = make_spider([0, 0.2], [0.5, 0.4], 0.03)
spider_infinite = make_spider_infinite([0, 0], 30, 0.03)

elt = make_elt_aperture(True)
gmt = make_gmt_aperture(True)
hicat = make_hicat_aperture(True)
tmt = make_tmt_aperture(True)
luvoir_a = make_luvoir_a_aperture(True)
luvoir_b = make_luvoir_b_aperture(True)
```
Current master branch:
```plain
HCIPy 0.4.1.dev196+geb74202 :
circle:     1.14 ms ± 11.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
rectangle:  591 µs ± 22.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
hexagon:    5.16 ms ± 145 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
spider:     9.01 ms ± 80.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
spider_inf: 5.68 ms ± 114 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
ELT:        604 ms ± 11.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
GMT:        178 ms ± 2.44 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
HiCAT:      84.6 ms ± 745 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
TMT:        392 ms ± 11.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
LUVOIR-A:   107 ms ± 902 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
LUVOIR-B:   42.1 ms ± 674 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
This PR:
```plain
HCIPy 0.4.1.dev200+g4395158 :
circle:     1.13 ms ± 10.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
rectangle:  583 µs ± 11.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
hexagon:    5.05 ms ± 162 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
spider:     1.87 ms ± 8.72 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
spider_inf: 1.78 ms ± 14.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
ELT:        150 ms ± 656 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
GMT:        178 ms ± 1.27 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
HiCAT:      28.6 ms ± 725 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
TMT:        103 ms ± 907 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
LUVOIR-A:   30.6 ms ± 341 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
LUVOIR-B:   12.4 ms ± 85.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```