@ehpor 2022-08-18
For the spiders, doing comparison in this way is actually faster than an `abs()`:

```python
import numpy as np
import numexpr as ne

x = np.random.randn(1024*1024)

def f(x):
    res = x < 1
    res *= x > -1
    
    return res

%timeit f(x)

def g(x):
    res = np.abs(x) < 1
    
    return res

%timeit g(x)

def h(x):
    res = x * x < 1
    
    return res

%timeit h(x)

def i(x):
    res = ne.evaluate("x * x < 1", local_dict={'x': x})
    
    return res

%timeit i(x)
```

gives

```
1.77 ms ± 103 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
3.57 ms ± 85.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
3.59 ms ± 32.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
630 µs ± 23.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

So doing the comparisons separately is actually twice as fast as the abs. Using numexpr is even faster, but considering future changes to multi-backend, it's not worth having specialized code for numpy just yet. That'll be for the future once backends are separated.