@ehpor 2022-08-18
> That is interesting. I wouldn't have thought that that would be the case. Okay no more comments from me then.

Yeah, me too at first. Something to do with the multiplication taking place in float64 vs boolean. I added a comment in the code with the counterintuitive result. I've used this trick too many times by now that it seemed obvious to me, but it's actually not trivial.

Merging after CI finishes.